feat: feat/add-server-init-presentation
