+++ presentations/dagger.json
@@ -6,6 +6,11 @@
   },
   "dependencies": [
     {
+      "name": "hugo",
+      "source": "github.com/stuttgart-things/dagger/hugo@v0.70.0",
+      "pin": "65455222342fd3546dbe4971859a967d48bed70d"
+    },
+    {
       "name": "templating",
       "source": "github.com/stuttgart-things/dagger/templating@v0.70.0",
       "pin": "65455222342fd3546dbe4971859a967d48bed70d"
