+++ tests/presentations/example-site/_index.md
@@ -1,0 +1,35 @@
++++
+title = "backstage"
+outputs = ["Reveal"]
+
+[logo]
+
+[reveal_hugo]
+history = true
+slide_number = true
+custom_theme = "reveal-hugo/themes/robot-lung.css"
+margin = 0.2
+mermaid = true
+highlight_theme = "color-brewer"
+transition = "slide"
+transition_speed = "fast"
+
+[reveal_hugo.templates.hotpink]
+class = "hotpink"
+background = "#FF4081"
++++
+
+<< slide id=agenda background-color="#A2D8FF" type="slide" transition="zoom" transition-speed="fast" background-image="https://artifacts.demo-infra.sthings-vsphere.labul.sva.de/images/stories2.png" background-size="500px" >>
+
+<% section %>
+
+<br/>
+<br/>
+<br/>
+<br/>
+<br/>
+<br/>
+
+🚀 Container stories 🚀<br/>🔁 OCI artifacts everywhere 🔁
+
+<% /section %>
