+++ tests/presentations/example-site/home/01-functions.md
@@ -1,0 +1,20 @@
++++
+weight = 20
++++
+
+{{< slide id=functions background-color="#B3B3FF" type="slide" transition="zoom" transition-speed="fast" >}}
+
+{{% section %}}
+
+# /FUNCTIONS
+
+single composable operations within a module
+
+<img src="https://i.kym-cdn.com/entries/icons/facebook/000/016/546/hidethepainharold.jpg" alt="Alt Text" width="800" style="border: 1px; box-shadow: none;" />
+
+```bash
+dagger call -m github.com/hidethepain@1.2.3 phone \
+--who=harold -vv --progress plain
+```
+
+{{% /section %}}
