+++ tests/presentations/example-site/hugo.toml
@@ -1,0 +1,16 @@
+baseURL = "/"
+languageCode = "en-us"
+title = "backstage"
+theme = ["github.com/joshed-io/reveal-hugo"]
+
+[module]
+  proxy = "direct"
+  vendored = true
+
+[markup.goldmark.renderer]
+unsafe = true
+
+[outputFormats.Reveal]
+baseName = "index"
+mediaType = "text/html"
+isHTML = true
