+++ tests/presentations/intro.md
@@ -1,0 +1,57 @@
++++
+weight = 10
++++
+
+{{< slide id=introduction background-color="#D4B9FF" type="slide" transition="zoom" transition-speed="fast" >}}
+
+{{% section %}}
+
+<img src="https://framerusercontent.com/images/dtoLxVnUgfKiPsuiXXWcTwxGIb4.png?lossless=1" width="750" style="border: none; box-shadow: none;" />
+
+**A tool to run composable workflows in containers**
+
+<br/>
+<br/>
+
+# /INTRODUCTION
+
+---
+
+### /AGENDA
+
+<img src="https://attachments-cdn-s.coub.com/coub_storage/coub/simple/cw_timeline_pic/7e14911f5a8/c1bec6bfed315483bad44/med_1587503875_image.jpg" alt="Alt Text" width="750"/>
+
+Intro • 🛠️ Functions • 💻 Architecture • 🔗 Modules 🔗 Integrations
+
+---
+
+> ### ❓ Audience
+>
+> **WHAT CI/CD DO YOU USE?**
+>
+> - 🦊 GitLab CI
+> - 🐙 GitHub Workflows
+> - ⚙️ Jenkins
+> - 🚀 Tekton
+> - ..?
+
+---
+
+### /Shift CI Left
+
+<img src="https://play.min.io/dagger/shiftci.jpg" alt="Alt Text" width="600" style="border: 1px; box-shadow: none;" />
+
+- Enable developers to test their code instantly - locally
+
+---
+
+> ### ❓ Audience
+>
+> **ARE YOU USING DAGGER (YET)?**
+>
+> - ✅ Yes, actively using it
+> - 🧪 Tested it briefly
+> - ❌ Not yet, but interested
+> - 🚫 No, not planning to use
+
+{{% /section %}}
